_initMap when element is loaded in a different location
If the element is opened in a second location, the _loadScript method will never call this._initMap() since window.google.maps already exists. Therefore, we need to init the `_map` in a different way (otherwise, the map won't load). Checking on firstUpdated(), with a check on `_map` and `window.google.maps` works.